# backend/internal/api/handler/auth_handler.go
package handler

import (
	"net/http"

	"github.com/gin-gonic/gin"

	"github.com/insmtx/Leros/backend/internal/adapter/account"
	"github.com/insmtx/Leros/backend/internal/api/contract"
	"github.com/insmtx/Leros/backend/internal/api/dto"
)

type AuthHandler struct {
	service account.AuthProvider
}

func NewAuthHandler(service account.AuthProvider) *AuthHandler {
	return &AuthHandler{service: service}
}

func (h *AuthHandler) RegisterRoutes(r gin.IRouter) {
	r.POST("/RegisterByEmail", h.RegisterByEmail)
	r.POST("/LoginByEmail", h.LoginByEmail)
	r.POST("/SendPhoneLoginCode", h.SendPhoneLoginCode)
	r.POST("/LoginByPhoneCode", h.LoginByPhoneCode)
	r.POST("/RefreshToken", h.RefreshToken)
	r.POST("/ChooseUin", h.ChooseUin)
	r.POST("/SwitchOrganization", h.SwitchOrganization)
	r.POST("/CreateOrganization", h.CreateOrganization)
	r.GET("/AuthSession", h.AuthSession)
}

func RegisterAuthRoutes(r gin.IRouter, service account.AuthProvider) {
	h := NewAuthHandler(service)
	h.RegisterRoutes(r)
}

// @Summary 邮箱注册
// @Description 使用邮箱和密码注册本地用户
// @Tags Auth
// @Accept json
// @Produce json
// @Param body body contract.RegisterByEmailRequest true "邮箱注册请求"
// @Success 200 {object} dto.Response "成功响应"
// @Failure 400 {object} dto.ErrorResponse "请求参数错误"
// @Failure 500 {object} dto.ErrorResponse "内部服务器错误"
// @Router /RegisterByEmail [post]
func (h *AuthHandler) RegisterByEmail(ctx *gin.Context) {
	var req contract.RegisterByEmailRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, dto.Error(dto.CodeInvalidParams, err.Error()))
		return
	}

	result, err := h.service.RegisterByEmail(ctx, &req.RegisterByEmailInput)
	if err != nil {
		handleAuthServiceError(ctx, err)
		return
	}
	ctx.JSON(http.StatusOK, dto.Success(result))
}

// @Summary 邮箱登录
// @Description 使用邮箱和密码登录并获取访问令牌
// @Tags Auth
// @Accept json
// @Produce json
// @Param body body contract.LoginByEmailRequest true "邮箱登录请求"
// @Success 200 {object} dto.Response "成功响应"
// @Failure 400 {object} dto.ErrorResponse "请求参数错误"
// @Failure 401 {object} dto.ErrorResponse "认证失败"
// @Failure 429 {object} dto.ErrorResponse "登录失败次数过多"
// @Failure 500 {object} dto.ErrorResponse "内部服务器错误"
// @Router /LoginByEmail [post]
func (h *AuthHandler) LoginByEmail(ctx *gin.Context) {
	var req contract.LoginByEmailRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, dto.Error(dto.CodeInvalidParams, err.Error()))
		return
	}

	result, err := h.service.LoginByEmail(ctx, &req.LoginByEmailInput)
	if err != nil {
		handleAuthServiceError(ctx, err)
		return
	}
	ctx.JSON(http.StatusOK, dto.Success(result))
}

// @Summary 发送手机登录验证码
// @Description 向手机号发送验证码；未配置短信服务时使用测试验证码
// @Tags Auth
// @Accept json
// @Produce json
// @Param body body contract.SendPhoneLoginCodeRequest true "发送手机验证码请求"
// @Success 200 {object} dto.Response "成功响应"
// @Failure 400 {object} dto.ErrorResponse "请求参数错误"
// @Failure 429 {object} dto.ErrorResponse "验证码发送太频繁"
// @Failure 500 {object} dto.ErrorResponse "内部服务器错误"
// @Router /SendPhoneLoginCode [post]
func (h *AuthHandler) SendPhoneLoginCode(ctx *gin.Context) {
	var req contract.SendPhoneLoginCodeRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, dto.Error(dto.CodeInvalidParams, err.Error()))
		return
	}

	result, err := h.service.SendPhoneLoginCode(ctx, &req.SendPhoneLoginCodeInput)
	if err != nil {
		handleAuthServiceError(ctx, err)
		return
	}
	ctx.JSON(http.StatusOK, dto.Success(result))
}

// @Summary 手机验证码登录
// @Description 使用手机号和验证码登录；首次登录会自动注册
// @Tags Auth
// @Accept json
// @Produce json
// @Param body body contract.LoginByPhoneCodeRequest true "手机验证码登录请求"
// @Success 200 {object} dto.Response "成功响应"
// @Failure 400 {object} dto.ErrorResponse "请求参数错误"
// @Failure 401 {object} dto.ErrorResponse "认证失败"
// @Failure 429 {object} dto.ErrorResponse "登录失败次数过多"
// @Failure 500 {object} dto.ErrorResponse "内部服务器错误"
// @Router /LoginByPhoneCode [post]
func (h *AuthHandler) LoginByPhoneCode(ctx *gin.Context) {
	var req contract.LoginByPhoneCodeRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, dto.Error(dto.CodeInvalidParams, err.Error()))
		return
	}

	result, err := h.service.LoginByPhoneCode(ctx, &req.LoginByPhoneCodeInput)
	if err != nil {
		handleAuthServiceError(ctx, err)
		return
	}
	ctx.JSON(http.StatusOK, dto.Success(result))
}

// @Summary 刷新令牌
// @Description 使用 refresh token 换取新的访问令牌
// @Tags Auth
// @Accept json
// @Produce json
// @Param body body contract.RefreshTokenRequest true "刷新令牌请求"
// @Success 200 {object} dto.Response "成功响应"
// @Failure 400 {object} dto.ErrorResponse "请求参数错误"
// @Failure 401 {object} dto.ErrorResponse "认证失败"
// @Failure 500 {object} dto.ErrorResponse "内部服务器错误"
// @Router /RefreshToken [post]
func (h *AuthHandler) RefreshToken(ctx *gin.Context) {
	var req contract.RefreshTokenRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, dto.Error(dto.CodeInvalidParams, err.Error()))
		return
	}

	result, err := h.service.RefreshToken(ctx, &req.RefreshTokenInput)
	if err != nil {
		handleAuthServiceError(ctx, err)
		return
	}
	ctx.JSON(http.StatusOK, dto.Success(result))
}

// @Summary 选择组织
// @Description 使用 refresh token 选择组织并获取访问令牌
// @Tags Auth
// @Accept json
// @Produce json
// @Param body body contract.ChooseUinRequest true "选择组织请求"
// @Success 200 {object} dto.Response "成功响应"
// @Failure 400 {object} dto.ErrorResponse "请求参数错误"
// @Failure 401 {object} dto.ErrorResponse "认证失败"
// @Failure 500 {object} dto.ErrorResponse "内部服务器错误"
// @Router /ChooseUin [post]
func (h *AuthHandler) ChooseUin(ctx *gin.Context) {
	var req contract.ChooseUinRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, dto.Error(dto.CodeInvalidParams, err.Error()))
		return
	}

	result, err := h.service.ChooseUin(ctx, &req.ChooseUinInput)
	if err != nil {
		handleAuthServiceError(ctx, err)
		return
	}
	ctx.JSON(http.StatusOK, dto.Success(result))
}

// @Summary 切换组织
// @Description 切换当前登录组织并获取新的访问令牌
// @Tags Auth
// @Accept json
// @Produce json
// @Param body body contract.SwitchOrganizationRequest true "切换组织请求"
// @Success 200 {object} dto.Response "成功响应"
// @Failure 400 {object} dto.ErrorResponse "请求参数错误"
// @Failure 401 {object} dto.ErrorResponse "认证失败"
// @Failure 500 {object} dto.ErrorResponse "内部服务器错误"
// @Router /SwitchOrganization [post]
func (h *AuthHandler) SwitchOrganization(ctx *gin.Context) {
	var req contract.SwitchOrganizationRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, dto.Error(dto.CodeInvalidParams, err.Error()))
		return
	}

	result, err := h.service.SwitchOrganization(ctx, &req.SwitchOrganizationInput)
	if err != nil {
		handleAuthServiceError(ctx, err)
		return
	}
	ctx.JSON(http.StatusOK, dto.Success(&contract.SwitchOrganizationResponse{JwtToken: result.JwtToken}))
}

// @Summary 创建组织并切换
// @Description 创建新组织/企业并切换当前登录组织
// @Tags Auth
// @Accept json
// @Produce json
// @Param body body contract.CreateOrganizationRequest true "创建组织请求"
// @Success 200 {object} dto.Response "成功响应"
// @Failure 400 {object} dto.ErrorResponse "请求参数错误"
// @Failure 401 {object} dto.ErrorResponse "认证失败"
// @Failure 500 {object} dto.ErrorResponse "内部服务器错误"
// @Router /CreateOrganization [post]
func (h *AuthHandler) CreateOrganization(ctx *gin.Context) {
	var req contract.CreateOrganizationRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, dto.Error(dto.CodeInvalidParams, err.Error()))
		return
	}

	result, err := h.service.CreateOrganization(ctx, &req.CreateOrganizationInput)
	if err != nil {
		handleAuthServiceError(ctx, err)
		return
	}
	ctx.JSON(http.StatusOK, dto.Success(result))
}

// @Summary 当前登录会话
// @Description 获取当前用户、当前组织和可切换组织列表
// @Tags Auth
// @Produce json
// @Success 200 {object} dto.Response "成功响应"
// @Failure 401 {object} dto.ErrorResponse "认证失败"
// @Failure 500 {object} dto.ErrorResponse "内部服务器错误"
// @Router /AuthSession [get]
func (h *AuthHandler) AuthSession(ctx *gin.Context) {
	result, err := h.service.AuthSession(ctx)
	if err != nil {
		handleAuthServiceError(ctx, err)
		return
	}
	ctx.JSON(http.StatusOK, dto.Success(result))
}

func handleAuthServiceError(ctx *gin.Context, err error) {
	errMsg := err.Error()

	switch {
	case isAuthBadRequestError(err):
		ctx.JSON(http.StatusBadRequest, dto.Error(dto.CodeInvalidParams, errMsg))
	case isAuthUnauthorizedError(err):
		ctx.JSON(http.StatusUnauthorized, dto.Error(dto.CodeInternalError, errMsg))
	case errMsg == "登录失败次数过多，请稍后再试":
		ctx.JSON(http.StatusTooManyRequests, dto.Error(dto.CodeInternalError, errMsg))
	case errMsg == "验证码发送太频繁，请稍后再试":
		ctx.JSON(http.StatusTooManyRequests, dto.Error(dto.CodeInternalError, errMsg))
	default:
		ctx.JSON(http.StatusInternalServerError, dto.Error(dto.CodeInternalError, errMsg))
	}
}

func isAuthBadRequestError(err error) bool {
	switch err.Error() {
	case "请输入邮箱",
		"请输入正确的邮箱",
		"请输入密码",
		"密码不一致",
		"密码长度不能少于8位",
		"密码长度不能超过20位",
		"密码不能包含中文",
		"密码不能包含空格",
		"8-20位，数字/大写字母/小写字母/字符至少3种",
		"该邮箱已注册",
		"请输入手机号",
		"请输入正确的手机号",
		"请输入验证码",
		"刷新令牌不能为空":
		return true
	case "组织名称不能为空":
		return true
	case "最多只能加入两个组织":
		return true
	case "account: 当前版本不支持此特性":
		return true
	case "account: 请先创建或加入组织":
		return true
	default:
		return false
	}
}

func isAuthUnauthorizedError(err error) bool {
	switch err.Error() {
	case "邮箱或密码错误",
		"验证码错误或已过期",
		"登录已过期，请重新登录",
		"用户不存在",
		"用户组织信息不存在",
		"用户未加入该组织",
		"请先登录":
		return true
	default:
		return false
	}
}